xat/pyproject.toml
[project]
name = "xat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "allure-pytest>=2.13.0",
    "sqlalchemy>=2.0.45",
    "python-jose>=1.0.0",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.21",
]
